services/web/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
flask-restful== 0.3.6
pyjwt==1.7.1
flask-cors==3.0.8